modify application services

+++ domain/services/payment_service.go
@@ -5,7 +5,7 @@ import "bytes"
 // PaymentService interface provides required methods
 // for a payment service
 type PaymentService interface {
-	GetTotalAmount(plate string) (float64, error)
-	Pay(plate string, amount float64) error
+	GetTotalAmount(plate string) (int, error)
+	Pay(plate string, amount int) error
 	PaymentReport() (*bytes.Buffer, error)
 }
